refactor: 코드리뷰 반영 (public, private)
