+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/converter/exception/ConverterExceptionControllerAdvice.java
@@ -1,8 +1,6 @@
 package com.startingblue.fourtooncookie.converter.exception;
 
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
